Updated the discussed changes.
Redefined and clarified the behavior of interrupt-level threshold (mintthresh)
- Clarified that the threshold is a CSR and not a memory mapped register. 
This is because using effective CSR-related atomic instructions can help 
reduce the length of critical sections.
- Redefined the threshold to be an 8-bit field holding only the setting 
of the associated privilege mode (and not of other modes). This definition
prevents unintended modification to thresholds in other modes.
- Clarified that interrupt threshold is only valid for the associated privilege mode.
This reduces the hardware cost because it only needs to select one global 
maximum and compare with the threshold of the associated mode.

Fixed the width of exccode in ucause and scause.
Fixed miscellaneous typos and mistakes.

Changed Interrupt Pending bits (clicintip) to be read-write to simplify the initialization process 
and make its definition cleaner (its attribute is no longer dependent on the operating modes).

Clarified software does not need to clear pending bits in the normal operation
because CLIC hardware already supports automatic clearing of pending bits for
edge-triggered interrupts.

Changed the definition of CLIC Base (mclicbase) from “read-only or WARL read-write” to just
“WARL read-write” for simplification.	

Removed the alignment restriction on the base address at mtvt.

Still need a read-only CSR address encoding for mintstatus (and s/h/u version). It was erroneously assigned to 
be at read-write address range before.