attempt at creating a AIA clic extension
Tried to separate CLIC spec into feature categories.  AIA already has a priority method so I removed the ability of clicintctl to represent a mix of level and priority.  clicintctl was renamed clicintlvl.  I also moved to the ideleg method of assigning privilege modes to interrupts.  With these changes, CLIC parameters was simplified and smclicconfig was rendered unnecessary.  Since clicintlvl only impacts the current privilege mode and CSRs mip/mie are mirrors of the lower 64 interrupts, there no longer appears that there is a need to require all priv modes are CLIC or all priv modes are CLINT.

Signed-off-by: smatherd <55101622+smatherd@users.noreply.github.com>