Post-review updates
there is one big question left to resolve: ACLIC + IMSIC configurations.

What's clear is that due to AIA CSRind space being a fixed window, if ACLIC and IMSIC were both present they would need to share the AIA-defined ei-prefixed registers.

But the way I read the AIA spec, the interrupt files are described as internal to the IMSIC and I guess this is how people will implement it.

IMO we should keep the original AIA eidelivery definition until we have a specification of how ACLIC and IMSIC could possibly be attached to the same hart.